+++ modules/cpu/render/pathtracer/PathTracer.cpp
@@ -34,6 +34,8 @@ void PathTracer::commit()
   getSh()->maxScatteringEvents = getParam<int>("maxScatteringEvents", 20);
   getSh()->maxRadiance = getParam<float>("maxContribution", inf);
   getSh()->numLightSamples = getParam<int>("lightSamples", -1);
+  getSh()->numFirstBounceLightSamples =
+      getParam<int>("firstBounceLightSamples", getSh()->numLightSamples);
 
   // Set shadow catcher plane
   const vec4f shadowCatcherPlane =